+++ src/onegov.core/onegov/core/tests/test_cronjobs.py
@@ -1,133 +1,19 @@
-import pytest
 import requests
 
 from datetime import datetime
 from freezegun import freeze_time
-from morepath.error import ConflictError
-from onegov.core import cronjobs, Framework
+from onegov.core import Framework
+from onegov.core.cronjobs import parse_cron, Job
 from onegov.core.utils import scan_morepath_modules
 from pytest_localserver.http import WSGIServer
-from sedate import ensure_timezone, replace_timezone
+from sedate import replace_timezone
 from sqlalchemy.ext.declarative import declarative_base
 from time import sleep
 from webtest import TestApp as Client
 
 
-def test_is_scheduled_at():
-    job = cronjobs.Job(lambda: None, hour=8, minute=0, timezone='CET')
-
-    assert job.hour == 8
-    assert job.minute == 0
-    assert job.timezone == ensure_timezone('CET')
-    assert job.name == 'test_is_scheduled_at.<locals>.<lambda>'
-
-    # dates must be timezone aware
-    assert job.is_scheduled_at(
-        replace_timezone(datetime(2015, 1, 1, 8), 'CET'))
-
-    # though we may also use this shortcut for the same thing
-    assert job.is_scheduled_at(datetime(2015, 1, 1, 8), 'CET')
-
-    # if there's no timezone we fail
-    with pytest.raises(AssertionError):
-        job.is_scheduled_at(datetime(2015, 1, 1, 8))
-
-    assert not job.is_scheduled_at(datetime(2015, 1, 1, 7, 59), 'CET')
-    assert job.is_scheduled_at(datetime(2015, 1, 1, 8, 0, 1), 'CET')
-    assert job.is_scheduled_at(datetime(2015, 1, 1, 8, 0, 59), 'CET')
-    assert not job.is_scheduled_at(datetime(2015, 1, 1, 8, 1), 'CET')
-
-    assert not job.is_scheduled_at(datetime(2015, 1, 1, 8, 0, 1), 'UTC')
-    assert not job.is_scheduled_at(datetime(2015, 1, 1, 8, 0, 59), 'UTC')
-    assert job.is_scheduled_at(datetime(2015, 1, 1, 7, 0, 1), 'UTC')
-    assert job.is_scheduled_at(datetime(2015, 1, 1, 7, 0, 59), 'UTC')
-
-    # hours can be wildcards
-    job = cronjobs.Job(lambda: None, hour='*', minute=0, timezone='CET')
-    assert job.is_scheduled_at(datetime(2015, 1, 1, 8, 0), 'UTC')
-    assert job.is_scheduled_at(datetime(2015, 1, 1, 9, 0), 'UTC')
-    assert not job.is_scheduled_at(datetime(2015, 1, 1, 9, 15), 'UTC')
-
-    job = cronjobs.Job(lambda: None, hour='*', minute=15, timezone='CET')
-    assert not job.is_scheduled_at(datetime(2015, 1, 1, 8, 0), 'UTC')
-    assert not job.is_scheduled_at(datetime(2015, 1, 1, 9, 0), 'UTC')
-    assert job.is_scheduled_at(datetime(2015, 1, 1, 9, 15), 'UTC')
-
-
-def test_overlapping_cronjobs():
-
-    class App(Framework):
-        pass
-
-    @App.cronjob(hour=8, minute=0, timezone='UTC')
-    def first_job(request):
-        pass
-
-    @App.cronjob(hour=8, minute=0, timezone='UTC')
-    def second_job(request):
-        pass
-
-    scan_morepath_modules(App)
-
-    with pytest.raises(ConflictError):
-        App.commit()
-
-
-def test_overlapping_cronjobs_with_wildcards():
-
-    class App(Framework):
-        pass
-
-    @App.cronjob(hour='*', minute=0, timezone='UTC')
-    def first_job(request):
-        pass
-
-    @App.cronjob(hour=5, minute=0, timezone='UTC')
-    def second_job(request):
-        pass
-
-    scan_morepath_modules(App)
-
-    with pytest.raises(ConflictError):
-        App.commit()
-
-
-def test_overlapping_cronjobs_with_wildcards_no_conflict():
-
-    class App(Framework):
-        pass
-
-    @App.cronjob(hour='*', minute=0, timezone='UTC')
-    def first_job(request):
-        pass
-
-    @App.cronjob(hour=5, minute=15, timezone='UTC')
-    def second_job(request):
-        pass
-
-    scan_morepath_modules(App)
-    App.commit()
-
-
-def test_non_5_minutes_cronjobs():
-
-    class App(Framework):
-        pass
-
-    @App.cronjob(hour=8, minute=1, timezone='UTC')
-    def first_job(request):
-        pass
-
-    scan_morepath_modules(App)
-
-    with pytest.raises(AssertionError):
-        App.commit()
-
-
-def test_cronjobs_integration(postgres_dsn, redis_url):
-
+def test_run_cronjob(postgres_dsn, redis_url):
     result = 0
-    cronjobs.CRONJOB_POLL_RESOLUTION = 1
 
     class App(Framework):
         pass
@@ -140,7 +26,7 @@ def test_cronjobs_integration(postgres_dsn, redis_url):
     def view_root(self, request):
         return {}
 
-    @App.cronjob(hour=8, minute=0, timezone='UTC')
+    @App.cronjob(hour='*', minute='*', timezone='UTC', once=True)
     def run_test_cronjob(request):
         nonlocal result
         result += 1
@@ -165,12 +51,18 @@ def test_cronjobs_integration(postgres_dsn, redis_url):
 
         with freeze_time(replace_timezone(datetime(2016, 1, 1, 8, 0), 'UTC')):
             requests.get(server.url)
-            sleep(2.5)
+
+            for i in range(0, 600):
+                if result == 0:
+                    sleep(0.1)
+                else:
+                    break
+
+            sleep(0.1)
+            assert result == 1
 
     finally:
         server.stop()
-
-    assert result == 1
 
 
 def test_disable_cronjobs(redis_url):
@@ -205,3 +97,125 @@ def test_disable_cronjobs(redis_url):
     client.get('/')
 
     assert not app.config.cronjob_registry.cronjob_threads
+
+
+def test_parse_cron():
+    assert tuple(parse_cron('*', 'hour')) == (
+        0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12,
+        13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23)
+
+    assert tuple(parse_cron('*/2', 'hour')) == (
+        0, 2, 4, 6, 8, 10, 12, 14, 16, 18, 20, 22)
+
+    assert tuple(parse_cron('5', 'hour')) == (5, )
+    assert tuple(parse_cron('*/20', 'minute')) == (0, 20, 40)
+
+
+def test_job_offset():
+    job = Job(test_job_offset, 8, 15, 'Europe/Zurich')
+    assert job.offset != 0
+
+    second_job = Job(test_job_offset, 8, 15, 'Europe/Zurich')
+    assert second_job.offset == job.offset
+
+    third_job = Job(lambda x: x, 8, 15, 'Europe/Zurich')
+    assert third_job.offset != job.offset
+
+
+def test_next_runtime():
+
+    def in_timezone(*args, timezone='Europe/Zurich'):
+        return replace_timezone(datetime(*args), timezone)
+
+    def next_runtime(hour, minute, timezone='Europe/Zurich'):
+        job = Job(lambda x: x, hour=hour, minute=minute, timezone=timezone)
+
+        # disable offset for tests
+        job.offset = 0
+
+        return job.next_runtime()
+
+    # fixed hour, minute
+    with freeze_time(in_timezone(2019, 1, 1, 8, 14, 59)):
+        assert next_runtime(hour=8, minute=15) \
+            == in_timezone(2019, 1, 1, 8, 15)
+
+    with freeze_time(in_timezone(2019, 1, 1, 8, 15, 0)):
+        assert next_runtime(hour=8, minute=15) \
+            == in_timezone(2019, 1, 1, 8, 15)
+
+    with freeze_time(in_timezone(2019, 1, 1, 8, 15, 1)):
+        assert next_runtime(hour=8, minute=15) \
+            == in_timezone(2019, 1, 2, 8, 15)
+
+    # any hour, fixed minute
+    with freeze_time(in_timezone(2019, 1, 1, 8, 14, 59)):
+        assert next_runtime(hour='*', minute=15) \
+            == in_timezone(2019, 1, 1, 8, 15)
+
+    with freeze_time(in_timezone(2019, 1, 1, 8, 15, 0)):
+        assert next_runtime(hour='*', minute=15) \
+            == in_timezone(2019, 1, 1, 8, 15)
+
+    with freeze_time(in_timezone(2019, 1, 1, 8, 15, 1)):
+        assert next_runtime(hour='*', minute=15) \
+            == in_timezone(2019, 1, 1, 9, 15)
+
+    # fixed hour, any minute
+    with freeze_time(in_timezone(2019, 1, 1, 8, 14, 59)):
+        assert next_runtime(hour=8, minute='*') \
+            == in_timezone(2019, 1, 1, 8, 15)
+
+    with freeze_time(in_timezone(2019, 1, 1, 8, 15, 0)):
+        assert next_runtime(hour=8, minute='*') \
+            == in_timezone(2019, 1, 1, 8, 15)
+
+    with freeze_time(in_timezone(2019, 1, 1, 8, 15, 1)):
+        assert next_runtime(hour=8, minute='*') \
+            == in_timezone(2019, 1, 1, 8, 16)
+
+    # any hour, every 15 minutes
+    with freeze_time(in_timezone(2019, 1, 1, 8, 14, 59)):
+        assert next_runtime(hour='*', minute='*/15') \
+            == in_timezone(2019, 1, 1, 8, 15)
+
+    with freeze_time(in_timezone(2019, 1, 1, 8, 15, 0)):
+        assert next_runtime(hour='*', minute='*/15') \
+            == in_timezone(2019, 1, 1, 8, 15)
+
+    with freeze_time(in_timezone(2019, 1, 1, 8, 15, 1)):
+        assert next_runtime(hour='*', minute='*/15') \
+            == in_timezone(2019, 1, 1, 8, 30)
+
+    with freeze_time(in_timezone(2019, 1, 1, 8, 45, 0)):
+        assert next_runtime(hour='*', minute='*/15') \
+            == in_timezone(2019, 1, 1, 8, 45)
+
+    with freeze_time(in_timezone(2019, 1, 1, 8, 45, 1)):
+        assert next_runtime(hour='*', minute='*/15') \
+            == in_timezone(2019, 1, 1, 9, 0)
+
+    # every 2 hours, every 15 minutes
+    with freeze_time(in_timezone(2019, 1, 1, 8, 14, 59)):
+        assert next_runtime(hour='*/2', minute='*/15') \
+            == in_timezone(2019, 1, 1, 8, 15)
+
+    with freeze_time(in_timezone(2019, 1, 1, 8, 15, 0)):
+        assert next_runtime(hour='*/2', minute='*/15') \
+            == in_timezone(2019, 1, 1, 8, 15)
+
+    with freeze_time(in_timezone(2019, 1, 1, 8, 15, 1)):
+        assert next_runtime(hour='*/2', minute='*/15') \
+            == in_timezone(2019, 1, 1, 8, 30)
+
+    with freeze_time(in_timezone(2019, 1, 1, 8, 45, 0)):
+        assert next_runtime(hour='*/2', minute='*/15') \
+            == in_timezone(2019, 1, 1, 8, 45)
+
+    with freeze_time(in_timezone(2019, 1, 1, 8, 45, 1)):
+        assert next_runtime(hour='*/2', minute='*/15') \
+            == in_timezone(2019, 1, 1, 10, 0)
+
+    with freeze_time(in_timezone(2019, 1, 1, 23, 59, 59)):
+        assert next_runtime(hour='*/2', minute='*/15') \
+            == in_timezone(2019, 1, 2, 0, 0)